src/dashboard-requirements.txt
ollama
pandas
streamlit
pydantic 
langchain_community
langchain_core
sqlalchemy 
psycopg2-binary 
python-dotenv
scikit-learn
langchain_ollama
tqdm
from_root
fastapi
uvicorn
langchain-google-genai